add classes for different log types
